requirements.txt
selenium
pytest
pytest-html
allure-pytest
requests
webdriver-manager
python-dotenv
jsonschema
faker
black
flake8
isort
